# glue/rexglue-sdk-main/community-server/api/openapi.yaml
openapi: 3.1.0
info:
  title: LibertyRecomp Community Service
  version: 2.0.0
  description: >-
    Community-operated compatibility API. This project is not affiliated with or authorized by
    Rockstar Games and never serves copyrighted game assets. Hex strings represent all Xbox
    64-bit identifiers and exchange keys. Binary values use RFC 4648 base64; auth uses base64url.
servers: [{url: "https://community.example.invalid"}]
security: [{bearerAuth: []}]
paths:
  /api/v2/devices/challenge:
    post:
      security: []
      operationId: createDeviceChallenge
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChallengeRequest'}, examples: {fixture: {value: {device_id: device_fixture, xuid: '0xe000000000000001', public_key: VlX3AZ7qGxATy5ZI4U6QR7YyJEfAbG-ZOr8kw-K0KYQ}}}}}}
      responses: {'201': {description: Challenge, content: {application/json: {schema: {$ref: '#/components/schemas/ChallengeResponse'}, examples: {fixture: {value: {challenge_id: ch_fixture, challenge: AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8}}}}}}}
  /api/v2/devices/enroll:
    post:
      security: []
      operationId: enrollOrVerifyDevice
      description: >-
        Enrolls a new public key or verifies the already-enrolled key. Sign the exact ASCII bytes
        of the base64url challenge. Fixture signature for the challenge/public key above is
        J6UVKhjVqhBrWJcy9XFhnfxpATK523cJWVEGFLWSiRMFdVSsVplWIWFSUYFJn2bq2gbGPveZG-DvIjVo-l-kBw.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EnrollRequest'}}}}
      responses: {'201': {description: Tokens, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}}
  /api/v2/devices/refresh:
    post:
      security: []
      operationId: refreshDeviceToken
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [refresh_token], properties: {refresh_token: {type: string}}}}}}
      responses: {'200': {description: Rotated tokens, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}}
  /api/v2/sessions:
    post:
      operationId: createSession
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SessionRecord'}}}}
      responses: {'201': {description: Created session, content: {application/json: {schema: {$ref: '#/components/schemas/SessionRecord'}}}}, '409': {$ref: '#/components/responses/Conflict'}}
  /api/v2/sessions/search:
    post:
      operationId: searchSessions
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SessionSearch'}}}}
      responses: {'200': {description: Matching sessions, content: {application/json: {schema: {type: object, required: [procedure_index,sessions], properties: {procedure_index: {type: integer}, sessions: {type: array, items: {$ref: '#/components/schemas/SessionRecord'}}}}}}}}
  /api/v2/sessions/{id}:
    parameters: [{$ref: '#/components/parameters/ResourceID'}]
    get: {operationId: getSession, responses: {'200': {description: Session, headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {$ref: '#/components/schemas/SessionRecord'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
    patch:
      operationId: patchSession
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SessionPatch'}}}}
      responses: {'200': {description: Updated session, content: {application/json: {schema: {$ref: '#/components/schemas/SessionRecord'}}}}, '409': {$ref: '#/components/responses/Conflict'}}
    delete:
      operationId: deleteSession
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/IfMatch'}]
      responses: {'200': {description: Deleted}}
  /api/v2/sessions/by-xbox-id/{xbox_id}:
    get:
      operationId: getSessionByXboxID
      parameters: [{name: xbox_id, in: path, required: true, schema: {$ref: '#/components/schemas/Hex64'}}]
      responses: {'200': {description: Full session record, content: {application/json: {schema: {$ref: '#/components/schemas/SessionRecord'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
...
  /api/v2/events:
    get:
      operationId: pollEvents
      description: Durable HTTP long-poll fallback. Resume with the last processed numeric event id.
      parameters:
        - {name: after, in: query, schema: {type: integer, format: int64, minimum: 0}}
        - {name: wait_ms, in: query, schema: {type: integer, minimum: 0, maximum: 30000}}
      responses: {'200': {description: Events, content: {application/json: {schema: {type: object, properties: {events: {type: array, items: {$ref: '#/components/schemas/Event'}}}}}}}}
  /api/v2/realtime:
    get:
      operationId: connectRealtime
      description: Upgrade to WebSocket with Authorization header. Each text frame is one Event.
      parameters: [{name: after, in: query, schema: {type: integer, format: int64, minimum: 0}}]
      responses: {'101': {description: Switching Protocols}}
  /api/v2/chat/messages:
    post:
      operationId: sendTextChatMessage
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TextChatInput'}}}}
      responses: {'201': {description: Message appended to each authenticated recipient event stream}}
  /api/v2/relay/routes:
    post:
      operationId: registerRelayRoute
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RelayRouteRequest'}}}}
      responses: {'200': {description: Registered route, content: {application/json: {schema: {$ref: '#/components/schemas/RelayRoute'}}}}}
    delete:
      operationId: unregisterRelayRoute
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RelayRouteRequest'}}}}
      responses: {'200': {description: Unregistered}}
  /api/v2/relay/datagrams:
    post:
      operationId: sendRelayDatagram
      description: Validates the complete batch before enqueueing and preserves input order per destination. A single Datagram remains accepted for compatibility. Encoded payload strings across one batch are limited to 1011712 bytes within the 1048576-byte HTTP body cap.
      requestBody: {required: true, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/Datagram'}, {$ref: '#/components/schemas/DatagramBatch'}]}}}}
      responses: {'202': {description: Entire batch queued, content: {application/json: {schema: {type: object, required: [accepted,accepted_count], properties: {accepted: {type: boolean}, accepted_count: {type: integer, minimum: 1, maximum: 64}}}}}}}
    get:
      operationId: receiveRelayDatagrams
      parameters:
        - {name: local_port, in: query, required: true, schema: {type: integer, minimum: 1, maximum: 65535}}
        - {name: max_bytes, in: query, required: true, schema: {type: integer, minimum: 1, maximum: 1048576}}
        - {name: wait_ms, in: query, required: true, schema: {type: integer, minimum: 0, maximum: 30000}}
      responses: {'200': {description: Datagrams, content: {application/json: {schema: {type: object, properties: {datagrams: {type: array, items: {$ref: '#/components/schemas/Datagram'}}}}}}}}
  /api/v2/matchmaking/tickets:
    post:
      operationId: createMatchmakingTicket
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MatchmakingTicket'}}}}
      responses: {'201': {description: Ticket, content: {application/json: {schema: {$ref: '#/components/schemas/MatchmakingTicket'}}}}}
  /api/v2/matchmaking/tickets/{id}:
    parameters: [{$ref: '#/components/parameters/ResourceID'}]
    get: {operationId: getMatchmakingTicket, responses: {'200': {description: Ticket, content: {application/json: {schema: {$ref: '#/components/schemas/MatchmakingTicket'}}}}}}
    delete: {operationId: cancelMatchmakingTicket, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/IfMatch'}], responses: {'200': {description: Cancelled}}}
...
  /api/v2/progression/{kind}:
    parameters: [{$ref: '#/components/parameters/ResourceKind'}]
    get: {operationId: listProgressionResources, responses: {'200': {description: Resources}}}
    post: {operationId: appendProgressionResource, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/ResourcePayload'}, responses: {'201': {description: Created}}}
  /api/v2/friends:
    get:
      operationId: enumerateFriends
      parameters: [{name: after, in: query, schema: {type: string}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}]
      responses: {'200': {description: XUID-ordered friend page, content: {application/json: {schema: {$ref: '#/components/schemas/FriendPage'}}}}}
  /api/v2/friends/check:
    post:
      operationId: checkFriends
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [xuids], properties: {xuids: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Hex64'}}}}}}}
      responses: {'200': {description: Input-ordered relationship flags}}
  /api/v2/friends/relationships:
    post:
      operationId: mutateFriendRelationship
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [target_xuid,action], properties: {target_xuid: {$ref: '#/components/schemas/Hex64'}, action: {type: string, enum: [request,accept,remove,block,unblock]}}}}}}
      responses: {'200': {description: Relationship projection}}
  /api/v2/invites:
    post:
      operationId: createInvites
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/InviteCreate'}}}}
      responses: {'201': {description: Created invitation records}}
    get:
      operationId: enumerateInvites
      parameters: [{name: after, in: query, schema: {type: string}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}, {name: state, in: query, schema: {type: string, enum: [pending,accepted,expired,revoked]}}]
      responses: {'200': {description: ID-ordered invitation page}}
  /api/v2/invites/{id}/accept:
    post:
      operationId: acceptInvite
      parameters: [{$ref: '#/components/parameters/ResourceID'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [expected_revision], properties: {expected_revision: {type: integer, format: int64, minimum: 1}}}}}}
      responses: {'200': {description: Accepted invitation plus full SessionRecord}}
  /api/v2/stats/writes:
    post:
      operationId: writeStats
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StatWrite'}}}}
      responses: {'200': {description: Idempotent write receipt}}
  /api/v2/stats/read:
    post:
      operationId: readStats
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [xuids,view_id,stat_ids], properties: {xuids: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Hex64'}}, view_id: {$ref: '#/components/schemas/Hex32'}, stat_ids: {type: array, maxItems: 64, items: {$ref: '#/components/schemas/Hex32'}}}}}}}
      responses: {'200': {description: Rows in input XUID order, content: {application/json: {schema: {type: object, required: [rows], properties: {rows: {type: array, items: {$ref: '#/components/schemas/StatRow'}}}}}}}}
  /api/v2/stats/reset:
    post:
      operationId: resetStatsView
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [view_id], properties: {view_id: {$ref: '#/components/schemas/Hex32'}}}}}}
      responses: {'200': {description: Authenticated caller row cleared in the requested view}}
  /api/v2/stats/skill:
    post:
      operationId: readSkill
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [xuids], properties: {xuids: {type: array, minItems: 1, maxItems: 100, items: {$ref: '#/components/schemas/Hex64'}}}}}}}
      responses: {'200': {description: Finite numeric ratings in exact input order, content: {application/json: {schema: {type: object, required: [ratings], properties: {ratings: {type: array, items: {type: object, required: [xuid,rating], properties: {xuid: {$ref: '#/components/schemas/Hex64'}, rating: {type: number}}}}}}}}}}
  /api/v2/leaderboards/{view_id}:
    get:
      operationId: enumerateLeaderboard
      parameters: [{name: view_id, in: path, required: true, schema: {$ref: '#/components/schemas/Hex32'}}, {name: stat_id, in: query, required: true, schema: {$ref: '#/components/schemas/Hex32'}}, {name: offset, in: query, schema: {type: integer, minimum: 0}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}, {name: friends_only, in: query, schema: {type: boolean}}]
      responses: {'200': {description: Descending rows sorted without integer precision loss; value is the typed StatValue object}}
...
components:
  securitySchemes: {bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$'}}
...
    ResourceID: {name: id, in: path, required: true, schema: {type: string}}
...
  responses:
    Conflict: {description: Revision, capacity, or state conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Hex64: {type: string, pattern: '^0x[0-9a-fA-F]{16}$', examples: ['0xe000000000000001']}
...
    ChallengeRequest: {type: object, additionalProperties: false, required: [device_id,xuid,public_key], properties: {device_id: {type: string}, xuid: {$ref: '#/components/schemas/Hex64'}, public_key: {type: string, description: base64url Ed25519 public key}}}
    ChallengeResponse: {type: object, additionalProperties: false, required: [challenge_id,challenge], properties: {challenge_id: {type: string}, challenge: {type: string, description: base64url challenge; sign its exact ASCII bytes}}}
    EnrollRequest: {type: object, additionalProperties: false, required: [challenge_id,device_id,xuid,machine_id,player_name,public_key,signature], properties: {challenge_id: {type: string}, device_id: {type: string}, xuid: {$ref: '#/components/schemas/Hex64'}, machine_id: {type: string}, player_name: {type: string, minLength: 1, maxLength: 32}, public_key: {type: string}, signature: {type: string}}}
    TokenPair: {type: object, additionalProperties: false, required: [access_token,refresh_token,expires_in], properties: {access_token: {type: string}, refresh_token: {type: string}, expires_in: {type: integer, format: int64}}}
...